PostgreSQL support: convert storage/duplicates/upgrades/cleanup, fix INTEGER overflow bug (64/70 files)
Converts services/storageForecast.ts, services/duplicates.ts,
services/upgradeCandidates.ts, and services/cleanupSuggestions.ts - the
first Tier 1 batch of a newly-written bottom-up conversion plan for the
remaining files (see DATABASE_MIGRATION.md). Fixed every downstream call
site across 3 still-unconverted files (system.ts, metrics.ts,
scheduler.ts) without converting those files themselves, including one
.map() callback needing the same Promise.all restructuring pattern used
in Rounds 93/95.

Found and fixed a serious, previously-latent Postgres bug unrelated to
this migration's own async conversion work: the schema translation
mapped SQLite's INTEGER (already 64-bit) to Postgres's INTEGER (32-bit,
max ~2.1GB) literally for 4 byte-count columns - queue.size,
recycle_bin.size_bytes, disk_usage_samples.free_bytes/total_bytes. Any
real download or file over ~2GB (a routine 4K remux) would silently
overflow on every Postgres deployment doing real downloads. Caught live
when disk-usage sampling failed against a real ~1TB test filesystem.

Fixed the fresh-install schema (BIGINT on all 4 columns) and added a
startup migration (ALTER COLUMN TYPE BIGINT, safe and idempotent) that
widens existing installs' columns on next restart with no data loss.

Verified live against a real Postgres container: simulated an existing
pre-fix install, confirmed the predicted "integer out of range" error,
then confirmed the startup migration fixes it and recovers correctly;
exercised disk usage sampling (a real ~1TB value), repeated-import
detection, upgrade-candidate detection, and both cleanup-suggestion
routes end to end - same sequence regression-checked against SQLite,
where this bug class doesn't exist at all.

+++ server/src/db/schema.postgres.sql
@@ -113,7 +113,7 @@ CREATE TABLE IF NOT EXISTS queue (
   indexer_id INTEGER REFERENCES indexers(id) ON DELETE SET NULL,
   download_client_id INTEGER REFERENCES download_clients(id) ON DELETE SET NULL,
   download_id TEXT,
-  size INTEGER,
+  size BIGINT,
   quality TEXT,
   status TEXT NOT NULL DEFAULT 'queued',
   progress REAL NOT NULL DEFAULT 0,
@@ -241,7 +241,7 @@ CREATE TABLE IF NOT EXISTS recycle_bin (
   title TEXT NOT NULL,
   original_path TEXT NOT NULL,
   recycle_path TEXT NOT NULL,
-  size_bytes INTEGER,
+  size_bytes BIGINT,
   deleted_at TEXT NOT NULL DEFAULT (to_char(now() AT TIME ZONE 'UTC', 'YYYY-MM-DD HH24:MI:SS')),
   restoring INTEGER NOT NULL DEFAULT 0, -- 1 while an async restore is in flight (large files move off the request thread)
   restore_error TEXT -- set if the last restore attempt failed, cleared on the next attempt
@@ -435,8 +435,8 @@ CREATE TABLE IF NOT EXISTS audit_log (
 CREATE TABLE IF NOT EXISTS disk_usage_samples (
   id SERIAL PRIMARY KEY,
   root_folder_id INTEGER NOT NULL REFERENCES root_folders(id) ON DELETE CASCADE,
-  free_bytes INTEGER NOT NULL,
-  total_bytes INTEGER NOT NULL,
+  free_bytes BIGINT NOT NULL,
+  total_bytes BIGINT NOT NULL,
   sampled_at TEXT NOT NULL DEFAULT (to_char(now() AT TIME ZONE 'UTC', 'YYYY-MM-DD HH24:MI:SS'))
 );
